requirements.txt
mysql-connector==2.1.4
